ui/fixtures/pyproject.toml
[project]
name = "fixtures"
version = "0.0.0"
readme = "README.md"
requires-python = ">=3.10"
# Unpin 'pandas' once they make a new release with musl-aarch64 wheels
dependencies = ["parquet-tools", "pandas==2.2.3", "pyarrow", "psycopg2-binary"]

[tool.uv]
constraint-dependencies = ["urllib3>=2.6.0"]
